more

+++ src/tls.zig
@@ -1,4 +1,4 @@
 const bun = @import("bun");
 
-pub var min_tls_version: ?f32 = null;
-pub var max_tls_version: ?f32 = null;
+pub var min_tls_version: ?f64 = null;
+pub var max_tls_version: ?f64 = null;
